+++ spec/features/admin_checks_student_loan_claim_spec.rb
@@ -8,7 +8,13 @@ RSpec.feature "Admin checking a Student Loans claim" do
   end
 
   scenario "service operator checks and approves a Student Loans claim" do
-    claim = create(:claim, :submitted, policy: StudentLoans)
+    claim = create(
+      :claim,
+      :submitted,
+      policy: StudentLoans,
+      student_loan_plan: StudentLoan::PLAN_2,
+      eligibility: build(:student_loans_eligibility, :eligible, student_loan_repayment_amount: "1987.65")
+    )
 
     visit admin_claims_path
     find("a[href='#{admin_claim_path(claim)}']").click
@@ -16,7 +22,8 @@ RSpec.feature "Admin checking a Student Loans claim" do
 
     expect(page).to have_content("1. Qualifications")
     expect(page).to have_content("2. Employment")
-    expect(page).to have_content("3. Decision")
+    expect(page).to have_content("3. Student loan amount")
+    expect(page).to have_content("4. Decision")
 
     click_on "Check qualification information"
     expect(page).to have_content("Qualifications")
@@ -30,6 +37,11 @@ RSpec.feature "Admin checking a Student Loans claim" do
     expect(page).to have_link(claim.eligibility.current_school.name)
 
     click_on "Complete employment check and continue"
+
+    expect(page).to have_content("Student loan amount")
+    expect(page).to have_content("£1,987.65")
+    expect(page).to have_content("Plan 2")
+    click_on "Complete student loan amount check and continue"
 
     expect(page).to have_content("Claim decision")